New Feature: Random Sensing matrices
Gaussian and Bernoulli sensing matrices added. Work reasonably well with dct. Work ok with Haar w
if scale is sufficiently high. The cascade does not work with them so I put a hack around it when using the random sensors, for now.
Need to figure out how to get the cascade working. Two ideas: 1) Use the RVM errors, like we did before. However, now we get no more exactly-zero errors. So perhaps something like 'discard measurement if error is too large'. However, we don't really trust the errors so I'm not sure how well this will work. This would make the cascade work for the DCT as well though. 2) Use the reconstructed values. If they are out of the [0.255] range, somehow discard the associated measurement. Although, we don't the associated measuremnent, do we?.
There are many differences in the code and it looks a lot less neat than before. Previous version remains untouched in the Code_V2 directory. New version  is in Code_V3/


Former-commit-id: 9bc4d44d42e77b3b20fe6c183f6464d02f6f8e10